Fix/silence shellcheck warnings.

+++ make_snapshot.sh
@@ -4,12 +4,14 @@
 # the public page at https://www.domjudge.org/snapshot/. Alternatively,
 # when a git URL is passed only a snapshot package is generated.
 
+# shellcheck disable=SC2086 # globally for PUBDIR, DJDIR, TEMPDIR
+
 set -e
 
 #DEBUG=1
 
 PUBDIR=~/public_html/snapshot
-DJDIR=domjudge-snapshot-`date +%Y%m%d`
+DJDIR=domjudge-snapshot-$(date +%Y%m%d)
 GITURL="https://github.com/DOMjudge/domjudge.git"
 
 # If a git repo URL is passed, don't update the website.
@@ -22,13 +24,13 @@ fi
 quiet()
 {
 	if [ "$DEBUG" ]; then
-		$@
+		"$@"
 	else
-		$@ > /dev/null 2>&1
+		"$@" > /dev/null 2>&1
 	fi
 }
 
-TEMPDIR=`mktemp -d /tmp/domjudge-make_snapshot-XXXXXX`
+TEMPDIR=$(mktemp -d /tmp/domjudge-make_snapshot-XXXXXX)
 cd $TEMPDIR
 
 git clone -q --no-checkout --depth 1 "$GITURL" dj-clone
@@ -36,14 +38,14 @@ git clone -q --no-checkout --depth 1 "$GITURL" dj-clone
 ( cd dj-clone && git archive --prefix=$DJDIR/ --format=tar refs/heads/main ) | tar x
 
 # Add released tag for revision information:
-sed -i "s/PUBLISHED =.*/PUBLISHED = `date +%Y-%m-%d`/" $DJDIR/paths.mk.in
+sed -i "s/PUBLISHED =.*/PUBLISHED = $(date +%Y-%m-%d)/" "$DJDIR/paths.mk.in"
 
 quiet make -C $DJDIR dist
 tar -cf $DJDIR.tar $DJDIR
 gzip -9 $DJDIR.tar
 
 if [ -n "$PUBDIR" ]; then
-	rm -rf $PUBDIR/*
+	rm -rf ${PUBDIR:?}/*
 	mkdir -p $PUBDIR/manual
 	cp -r $DJDIR/doc/manual/build/html/*                $PUBDIR/manual/
 	cp $DJDIR/doc/manual/build/domjudge-team-manual.pdf $PUBDIR/
@@ -51,6 +53,6 @@ if [ -n "$PUBDIR" ]; then
 	cd /
 fi
 
-[ "$DEBUG" ] || rm -rf $TEMPDIR
+[ "$DEBUG" ] || rm -rf "$TEMPDIR"
 
 exit 0
